Fix GH-13764: xsl cannot build on PHP 8.4
Move some of the DOM APIs from the non-public php_dom.h header to the
public header xml_common.h.
