+++ ext/dom/xml_common.h
@@ -80,4 +80,45 @@ PHP_DOM_EXPORT xmlNodePtr dom_object_get_node(dom_object *obj);
 	__id = ZEND_THIS; \
 	DOM_GET_OBJ(__ptr, __id, __prtype, __intern);
 
+struct _php_dom_libxml_ns_mapper;
+typedef struct _php_dom_libxml_ns_mapper php_dom_libxml_ns_mapper;
+
+static zend_always_inline php_dom_libxml_ns_mapper *php_dom_get_ns_mapper(dom_object *intern)
+{
+	ZEND_ASSERT(intern->document != NULL);
+	return (php_dom_libxml_ns_mapper *) intern->document->private_data;
+}
+
+static zend_always_inline xmlNodePtr php_dom_next_in_tree_order(const xmlNode *nodep, const xmlNode *basep)
+{
+	if (nodep->next) {
+		return nodep->next;
+	} else {
+		/* Go upwards, until we find a parent node with a next sibling, or until we hit the base. */
+		do {
+			nodep = nodep->parent;
+			if (nodep == basep) {
+				return NULL;
+			}
+			/* This shouldn't happen, unless there's an invalidation bug somewhere. */
+			if (UNEXPECTED(nodep == NULL)) {
+				zend_throw_error(NULL, "Current node in traversal is not in the document. Please report this as a bug in php-src.");
+				return NULL;
+			}
+		} while (nodep->next == NULL);
+		return nodep->next;
+	}
+}
+
+static zend_always_inline bool php_dom_follow_spec_doc_ref(const php_libxml_ref_obj *document)
+{
+	return document != NULL && document->class_type == PHP_LIBXML_CLASS_MODERN;
+}
+
+static zend_always_inline bool php_dom_follow_spec_intern(const dom_object *intern)
+{
+	ZEND_ASSERT(intern != NULL);
+	return php_dom_follow_spec_doc_ref(intern->document);
+}
+
 #endif
